+++ app/Http/Controllers/UserController.php
@@ -7,6 +7,8 @@ use App\Http\Requests\UpdateUserRequest;
 use App\Models\User;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Redirect;
 use Inertia\Inertia;
 
@@ -26,30 +28,40 @@ class UserController extends Controller
             ->paginate(5)
             ->withQueryString();
         return Inertia::render('User/Index', [
-            'users' => $users
+            'users' => $users,
+            'can' => [
+                'create' => Auth::user()->is_admin === 1,
+                'update' => Auth::user()->is_admin === 1,
+                'delete' => Auth::user()->is_admin === 1
+            ]
         ]);
     }
     public function create () {
+        Gate::authorize('create', User::class);
         return Inertia::render('User/Create');
     }
 
     public function store (StoreUserRequest $request) {
+        Gate::authorize('create', User::class);
         User::create($request->validated());
         return Redirect::route('users.index');
     }
 
     public function edit (User $user) {
+        Gate::authorize('update', $user);
         return Inertia::render('User/Edit', [
             'user' => $user
         ]);
     }
 
     public function update (UpdateUserRequest $request, User $user) {
+        Gate::authorize('update', $user);
         $user->update($request->validated());
         return Redirect::route('users.index');
     }
 
     public function destroy (User $user) {
+        Gate::authorize('delete', $user);
         $user->delete();
         return Redirect::route('users.index');
     }
